change volume and number of chapter type

+++ src/KamiYomu.CrawlerAgents.Core/Catalog/Builders/PageBuilder.cs
@@ -25,7 +25,7 @@
             return this;
         }
 
-        public PageBuilder WithPageNumber(int pageNumber)
+        public PageBuilder WithPageNumber(decimal pageNumber)
         {
             _page.PageNumber = pageNumber;
             return this;
